refactor(scaffold): move templates/app/_utils to templates/lib/utils + write cn to lib/utils/cn.ts
Mirror of the blog refactor in the scaffold templates. app/ is for
routing-convention files only; the per-app browser-safe helpers go
under lib/utils/.

Template moves:
  templates/app/_utils/ui.ts -> templates/lib/utils/ui.ts
  (templates/app/ dir removed; the scaffold has no template files
  in app/ for the full-stack/saas branches anymore.)

create.js changes:
  - utilsDir destination switches from app/_utils to lib/utils.
  - writeUiBootstrap writes cn() to lib/utils/cn.ts (not lib/utils.ts).
  - readUiComponent rewrites '../lib/utils.ts' -> '../../lib/utils/cn.ts'
    when copying registry components into components/ui/.
  - components.json aliases.utils becomes 'lib/utils/cn' so future
    'webjs ui add' resolves correctly via get-config.js's +'.ts' suffix.
  - The scaffolded example page.ts imports from '../lib/utils/ui.ts'.
  - Success-message file tree reflects the new structure.

Test updates assert lib/utils/cn.ts and lib/utils/ui.ts existence,
the new aliases.utils value, and the rewritten button.ts import.
